Damage position, closes #82

+++ abserv/abserv/SkillsComp.cpp
@@ -96,7 +96,7 @@ void SkillsComp::Write(Net::NetworkMessage& message)
         {
             message.AddByte(AB::GameProtocol::GameObjectUseSkill);
             message.Add<uint32_t>(owner_.id_);
-            message.Add<int>(lastSkillIndex_ + 1);
+            message.Add<uint8_t>(static_cast<uint8_t>(lastSkillIndex_ + 1));
             if (auto ls = lastSkill_.lock())
             {
                 message.Add<uint16_t>(static_cast<uint16_t>(ls->GetRealEnergy()));
@@ -118,7 +118,7 @@ void SkillsComp::Write(Net::NetworkMessage& message)
         {
             message.AddByte(AB::GameProtocol::GameObjectSkillFailure);
             message.Add<uint32_t>(owner_.id_);
-            message.Add<int16_t>(static_cast<uint16_t>(lastSkillIndex_ + 1));
+            message.Add<int8_t>(static_cast<uint8_t>(lastSkillIndex_ + 1));
             message.AddByte(lastError_);
         }
         startDirty_ = false;
@@ -130,14 +130,14 @@ void SkillsComp::Write(Net::NetworkMessage& message)
         {
             message.AddByte(AB::GameProtocol::GameObjectEndUseSkill);
             message.Add<uint32_t>(owner_.id_);
-            message.Add<int>(lastSkillIndex_ + 1);
+            message.Add<int8_t>(static_cast<uint8_t>(lastSkillIndex_ + 1));
             message.Add<uint16_t>(static_cast<uint16_t>(newRecharge_));
         }
         else
         {
             message.AddByte(AB::GameProtocol::GameObjectSkillFailure);
             message.Add<uint32_t>(owner_.id_);
-            message.Add<int16_t>(static_cast<uint16_t>(lastSkillIndex_ + 1));
+            message.Add<int8_t>(static_cast<uint8_t>(lastSkillIndex_ + 1));
             message.AddByte(lastError_);
         }
         endDirty_ = false;
